[Refactor] 작업 지시서 화면 구조와 유틸 로직 분리
